pixel-arty/pyproject.toml
[project]
name = "pixel-arty"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
	"torch",
	"torchvision",
	"matplotlib",
	"numpy",
	"pillow",
	"dotenv"
]
